mainのマージ

+++ src/test/scala/b4processor/utils/ExecutorValue.scala
@@ -1,3 +1,3 @@
 package b4processor.utils
 
-case class ExecutorValue(destinationTag: Int, value: Int)
+case class ExecutorValue(destinationTag: Int, value: BigInt)
